backend/requirements.txt
fastapi
uvicorn
openai
qdrant-client
python-dotenv
langchain
langchain-community
langchain-openai
langchain-qdrant
langchain-text-splitters